requirements.txt
pandas-gbq
google-auth
requests
python-dotenv